refactor(cli): split pack command installer resolution and upload flow
