+++ docs/architecture/maintainability-baseline.txt
@@ -4,7 +4,6 @@
 1013 crates/surge-bench/src/runner.rs
 813 crates/surge-cli/src/commands/install/mod.rs
 1860 crates/surge-cli/src/commands/install/remote.rs
-1189 crates/surge-cli/src/commands/pack.rs
 835 crates/surge-cli/src/main.rs
 861 crates/surge-core/src/config/manifest.rs
 727 crates/surge-core/src/install.rs
